Remove unhelpful FK indexes
The removed indexes are on FK fields which take one of only a few
values, making them useless.  This lead to very slow performance if
SQLite chooses a bad index.

For instance, if SQLite picks the index on snomedct_description.type_id
in the query below, it ends up doing a scan of most of the
snomedct_description table in order to find a single result.

	SELECT *
	FROM snomedct_description
	WHERE concept_id = '900000000000451002'
	  AND type_id = '900000000000003001';