Switch CharField for TextField
SQLite doesn't enforce limits on varchars [1] so our max_lengths were
being ignored.  Under the hood SQLite stores varchars as TEXT, the same
as Postgres.  Since neither database is using a special storage type for
varchars this converts all our CharField instances to TextField.

This fixes the snomed fixture not loading when using Postgres because
the Description.effective_time field was set to 8 characters but needed
10.

1: https://www.sqlite.org/faq.html#q9