Stop using History Substitution Table to generate SNOMED CT codelists
Previously, the algorithm to generate a SNOMED CT codelist from a CTV3
codelist was:

* take list of CTV3 concepts
* use assured ctv3sctmap2 mappings to get active SNOMED CT concepts
* use Query Table to get any inactive SNOMED CT concepts that have been
  replaced by these active SNOMED CT concepts
* use History Substitution Table (HST) to get extra active concepts that
  replace these inactive SNOMED CT concepts

However, the HST contains some substitutions that are clearly wrong.
For instance there is a substitution from the old concept:

  [X]Malignant neoplasm of overlapping lesion of bone and articular cartilage (disorder)

to the new concept:

  Clinical finding (finding)

I've spent a bit of time trying to identify whether there is any extra
metadata in the HST that identifies which substitutions are more ikely
than others to be correct, but haven't got very far.

The trade-off is between:

(1) use the HST and incorrectly include some concepts
(2) don't use the HST and incorrectly exclude some concepts

The risks of (1) are that we end up with a very high level concept like
Clinical finding in a SNOMED CT codelist.  This will render the codelist
builder (which will be used to review the automatically generated
codelists) unusable.

The risks of (2) are that we miss out concepts that should be included
in a codelist.  When reviewing, it is going to be easier to exclude some
incorrectly included concepts than it is to include some incorrectly
excluded concepts that you can't see.

However, any incorrectly excluded concepts are very likely to either:

* be descendants of the active concepts which are directly mapped from
  the CTV3 concepts
* have names that are similar to those active concepts

In both cases, they'll be visible in the codelist builder.

So the risks of (2) are minimal, and this is what I've decided to go
with.

There is an option (3) which is to spend more time trying to understand
the documentation, but I think we should be pragmatic about this.