Revert "Tell Fabric to explicitly use the user's ssh config"
This reverts commit d83b469cc24861149b4884343ca775f564b5d3c7.

Disable fabric's use_ssh_config option so automated deploys by EBMBot
can work.  With this set to true EBMBot deploys require a password on
the server.

This was added to enable deploys from macOS by @ghickman, however we
later discovered the issue was macOS not adding SSH keys to the agent by
default.