Restrict redirect-checking tests to only check redirects
These tests check the redirects of codelist() and version() fire as
expected.  However they also called the full view previously which meant
any bug in the rest of the view failed these tests, making them brittle.
This restricts those tests to only check the redirect fired, that it
fired correctly, decoupling them from the rest of the view logic.