Extract grouping codes by type from tree_tables
CTV3 codes can be in more than one group (and we want some extra control
over that due to "Additional values" being a bit spurious) so we want
the coding system modules to handle that for us.  This pulls the
grouping of codes out of tree_tables and pushes it into the CTV3&Snomed
coding_system modules, removing their code_to_type functions (only used
as a sort key in tree_tables) in the process.