Fix dependency on env.app
Openprescribing has multiple deployments (prod and staging).  The
`setup_sudo()` function was copied from OP's fabfile.  OP's fabfile
differentiates between the two environments using env.app, but we don't
have that distinction in opencodelists.  This removes the use of env.app
but keeps the same format as OP in case we switch to a two env
deployment model.