Add string-foot-gun guard to Hierarchy.from_codes
Hierarchy.from_codes will accept any type of iterable for its `codes`
argument.  However if that iterable is a string it will produce a
Hierarchy instance with no nodes.  The string is passed to the given
`coding_system`'s ancestor/descendant relationships functions which
cast it to a list.  In the case of a string this breaks it down into
its contained characters, which doesn't return any [correct] results
from the database.