Add CodelistLabel model
This will allow a codelist to move between projects and change names,
while allowing us to keep the old URL working.

Note that the data migration has no backwards step and is not tested,
and the tests for a previous data migration have been deleted (because
it failed, since the new data migration has no backwards step).