Rewrite the tree-rendering code for tree tab
If a concept's descendants all have the same state as the descendant,
then when the tree page is first shown, the concept's descendants are
now collapsed.  This means that large codelists are now renderable.

Previously the whole tree was expanded, and because many SNOMED concepts
have many parents, some concepts would appear hundreds or thousands of
times in the trees of large codelists.  React would struggle to render
this, and in any case it was unwieldy for a user.

Further performance improvements are possible by using PureComponent
and/or shouldComponentUpdate, but this is Good Enough for now.

Additionally, this new approach changes what happens when a user clicks
on the ⊞ or ⊟ to toggle whether a concept's children are visible.
Previously, this would toggle the visibility of the concept's children
for each occurrence of the concept in the tree.  This led to surprising
behaviour where clicking on the toggle would move the concept up or down
the page if there was another occurrence of the concept higher up the
page.  Now only the children that occur below the occurrence of the
concept that has been clicked will be toggled.  This is achieved by
being explicit about paths through the tree.

There are a couple of other behaviour changes:

* The colours have changed.  Included concepts are black, excluded
  concepts are grey.  This matches the colours used in the builder.
* There are no links to the browser.  These could be added back later
  when the browser shows something more useful.

This PR creates some duplication, and leaves some unused code lying
around.  I plan to rewrite the builder to use the same approach (it also
suffers from performance problems with large codelists) and once I have
done so I will tidy things up.