Add extra fixture
This serves as a regression test for the fix in a8129728.

Specifically, the bug that a8129728 fixes was triggered by the presence
of ancestor codes that are excluded with all their descendants.  By
creating a search for "tennis" and excluding "tennis toe", "tennis toe"
is an ancestor code that would trigger this bug.