Bump gunicorn timeout
For codelists with a lot of searches, exporting the codelist to the
buidler to be edited takes longer than 30s.  This is because we repeat
each search and this can be slow.

A better fix would be add FTS indexes to speed up searching, but for now
bumping the gunicorn timeout will have to do.