Change how searching by code works
Instead of having two buttons, one for searching by code and one for
searching by term (which seems to be causing confusion) we now have a
single "search" button.  By default, we search by term, but if the
search string matches "code:xxx", we search for code xxx instead, and if
the search string matches "code:xxx*", we search for concepts whose code
begins with xxx.