Only render TreeTables if there are changes
Rendering a TreeTables component for a codelist with 5k codes takes
around 2s.  Including or excluding a code triggers three re-renders.
Rather than work out how to streamline this (which would be sensible),
we check for changes to codeToStatus and visiblePaths, and only render
TreeTables if these have changed.

Responding to an inclusion or exclusion now takes ~2.5s instead of
~6.5s.

I have investigated adding shouldComponentUpdate methods to TreeTable
and Tree, but these do not bring appreciable gains.  I suspect this is
because the work to determine whether a TreeTable or a Tree should be
rendered is almost as involved as rendering it.

To speed it up further, we could investigate window functions.  Apart
from that, I'm out of ideas, unless we find a way to indicate to
components that they should update, which seems to negate most of the
benefits of React.