Bump black and remove black exclusions from pyproject.toml
Black's behaviour has changed so that .gitignore is ignored if it's
given any explicit exclusions.

This means that the latest dependabot PR is failing, because it's
checking the venv directory in CI.

The fix is to remove the exclusions from pyproject.toml, and to run
black over the now-included files.  These are mostly Django migrations,
with a few autogenerated ANTLR helper files.