Properly enforce uniqueness constraint for handle slugs/names
A handle is an object that links a codelist to an owner (a user or an
organisation), a name, and a slug.  This is to allow us to change a
codelist's owner or rename/reslug it, without breaking existing URLs.

We want to enforce that an owner cannot have two handles with the same
name or slug.  However, we have been doing this incorrectly.  This means
that there are a handful of duplicate handles.

To fix, we first have to handle the duplicate handles.  This is done in
migration 0045, which updates the names/slugs of duplicate handles.
This migration has no tests, but I have verified that it changes only
the handles that I expect.

Then we can fix the unique_together constraint on Handle.  Migration
0046 adds this to the database, and this migration fails unless 0045 has
been run, which gives us added confidence that 0045 is correct.

Fixes #866