Bump slug field sizes
The slugs are determined from the names, so the sizes of the fields
should match!

This fixes a problem where you couldn't update the metadata of a
codelist with a long name, because the browser would refuse to submit a
form with a long slug.

Changing the size of the search slug is a precautionary measure.