Handle quotes in CSV files as expected
We turned off quote handling in CSV uploading to handle confusing cases
where headers contained extraneous whitespace (see #1065).  However,
that meant that we could not handle CSVs with commas (#1210).

Instead of attempting to handle CSVs with whitespace in headers, we now
raise a ValidationError, which allows us to to handle quotes properly.

Fixes #1210.