Display dm+d terms from CSV data for unknown codes in diffs
If we look up a term in the codelist version's coding system release,
and it doesn't exist, it will return the term "Unknown".  For dm+d
codelists, we will always have the original CSV data as uploaded, so
where a code is unknown, and it's available in the CSV data, we can
show that.