Management command to manually create and fix a draft codelist version
If the coding system has changed since a CodelistVersion was created, it
may not be possible to create a new version from the builder frontend.
Specifically, the export_to_builder function reruns the searches from
then original codelist version, and will fail if it encounters any
codes with unknown status (this happens if a search brings back new
codes that were not present on the original codelist version).

Currently we have the mechanisms available to generate and fix up a
new draft, but we choose not to do so, in order to avoid automatically
assigning statuses to new codes on a codelist version without the
user being made very explicitly aware of them.

This management command allows manual creation of a new draft (ignoring
unknown codes) and then immediately updating it to assign implicit
statuses to unknown codes where we can. It will output the codes that
have had their status updated; this can be be reported to the user so
they can check and confirm them.