Update export_to_builder to return unknown code statuses
If a new draft is created after the coding system has changed, the
coding system may contain new codes. If any of those codes are
descendants of any code on the draft's, and do not have corresponding
CodeObjs on the draft, the builder frontend will not be able to render them.
However, it CAN deal with CodeObjs with status ?, which will be displayed as
unresolved for the user to deal with.