Make dmd imports more memory efficient
dm+d imports have been failing since 10th July. There is no specific error,
and it succeeds when run locally, but the process is killed at some point when
run in production on dokku3. When I ran it manually and watched htop, it looks
like it's maxing out memory.

This is an attempt to decrease memory usage by using generators and etree's iter
methods wherever possible, and avoiding turning the entire loaded data files into
lists. I used tracemelloc.get_traced_memory() as a basic comparison (wrapped around
the import_coding_system() call), and ran the entire import with the old and new
code. Reported peak memory for the old code was 130,356,648, and for the
new code (this commit) it was 186,501.

The database created from both traced runs is identical in size and counts for dmd
models are the same, and all import data tests are passing.