Add a test to ensure CodingSystemRelease.update_or_create behaves correctly
As of Django 4.2, update_or_create passes update_fields
to Model.save()
This means that any fields that are updated in a custom
save() method may need to be added to update_fields in order to
ensure that they are also saved
https://docs.djangoproject.com/en/4.2/releases/4.2/#setting-update-fields-in-model-save-may-now-be-required

This test verifies that the validation checks in the
CodingSystemRelease save() method mean we don't need to
add extra fields to update_fields.

Other models with custom save methods that could be affected
(User, Organisation) add a slugified primary key field only if
it's None, which can only happen on an initial save, so won't
be affected by this.