Respect fixed_headers for non-dm+d codelist downloads
For dm+d codelists, we now default to including mapped VMPs
in CSV downloads. In order to do that, we need to use the code that
downloads with fixed headers (i.e. two columns, headed 'code' and
'term'), so for dm+d codelists ONLY, we need toignore the default value of
fixed_headers (False).