Add an API endpoint to check codelists from study repos
This endpoint may be called from opensafely-cli or from job-server.
It takes as POST parameters the content of a study repo's
codelists.txt file (specifying the codelists that are to be used in
the study), and a manifest (codelists.json) file, which is generated
when `opensafely codelists update` is run. It checks that the hashed
codelist data in the manifest file matches the content of the current
codelist versions, if they were to be downloaded again.

Note opensafely-cli already has a codelists check command, which checks
that the manifest file is consistent with the codelists in the study
repo. However, dm+d codelists can change due to new VMP mappings, which
get mapped into CSV downloads. Therefore a dm+d codelist that was
downloaded into a study repo may need to be updated again, even if no changes
have been made to the codelists.csv.