Fix the regex used to match codelists in codelists_check
We need to match codelists by tag or hash; tag doesn't go through
any cleaning, so it can techincally be any string, including random
characters and forward slashes. We're idetifying whether a codelist
in a codelists.txt file is valid by using forward slashes as a
delimiter between user-or-org/codelist-slug/tag-or-hash, so we need
to exclude forward slashes from the allowed characters for the
tag-or-hash part of the regex. This is fine, since any codelist
version that was created with forward slashes in the tag would have
a broken url anyway.