Check against shas of alternative ok downloads
Recent changes to the default format of CSV downloads mean
that a dm+d CSV downloaded now will be in a different format
to one downloaded a few weeks ago, even if the codes have not
changed. In order to minimise the impact on users, we can
check the shas of CSV downloads in the study repo manifest files
against possible downloaded versions that include all the
expected codes, even if they're not in the exact current download
format.