Disable the download-with-vmp button if a dm+d codelist is undownloadable
By undownloadable, we mean that it doesn't have an identifiable code
column header in the csv_data field, so it's undownloadable for OSI, and
also can't be used to retrieve mapped VMPs. We now validate codelists created
by csv upload, so this only applies to particularly old codelists. These
can still be downloaded in their original format.

e.g. https://opencodelists.org/codelist/opensafely/methotrexate-oral/738b5e12/#full-list
contains codes in a "VMP" column, which is not one of the allowed code headers.