Ensure prod smoketest can create/read db file
In actual prod, we mount a volume that is owned by the non-root
docker user. In the prod smoketest, we need to make sure that the
volume mounted in is also owned by the non-root user. The db doesn't
exist at this stage, so it will be created in the test.