Set `fetch-depth` for `actions/checkout`
This is to attempt to solve failures to create a new Sentry release of
the form:

```
Error: Command failed: /action-release/node_modules/@sentry/cli-linux-x64/bin/sentry-cli releases set-commits 14e3263812a1c9b8e84a014436488ecfdf86f131 --auto
error: Could not find the SHA of the previous release in the git history. If you limit the clone depth, try to increase it. Otherwise, it means that the commit we are looking for was amended or squashed and cannot be retrieved. Use --ignore-missing flag to skip it and create a new release with the default commits count.

Add --log-level=[info|debug] or export SENTRY_LOG_LEVEL=[info|debug] to see more output.
Please attach the full debug log to all bug reports.
```