Increase timeout to 45s
Several users have experienced `OperationalError`s because the database
has been locked. Whilst these errors relate to the builder, it's proving
hard to recreate them locally. It's not the case, for example, that
making many updates to a single codelist in a short space of time locks
the database.

Finding the route cause of the errors will take time. However, given
that the error has occurred 55 times in the last 30 days, increasing the
timeout from 30s to 45s and monitoring the errors seems reasonable.

Thanks to @emprestige for reporting the errors in the first instance.

Closes #1811