Replace pysqlite3-binary with sqlean.py
The advantage of sqlean.py is that it includes wheels for Windows,
macOS, and Linux. pysqlite3-binary doesn't: it includes wheels only for
Linux, and, based on the project's closed issues, that is the way it's
going to stay. (Aside: sqlean.py's maintainer has dropped into
pysqlite3-binary's issues to recommend sqlean.py over pysqlite3-binary
[1][].)

As we've made the switch in ehrQL, and as at least one member of the
team doesn't develop on Linux (and has only so much patience for
Docker), it would seem reasonable to replace the former for the latter.

[1]: https://github.com/coleifer/pysqlite3/issues/64