Add updated date to codelist version template
The codelist versions are rendered in date creation order, but this
is not always obvious to users.
Even then, we do not surface the creation of modification date anywhere.

For a user, trying to find a version of a codelist at a particular
point in time, this change provides the neccesary information.