Move the web health check to `app.json`
Mostly, we port the web health check from `CHECKS` to `app.json`. We set
the `Host` header to prevent Django from raising a `DisallowedHost`
error.

Why would Django raise this error? The web health check is performed by
invoking `curl` from the Dokku host; an internal IP is passed as the
URL,[^1] which means that Cloudflare DNS and nginx are bypassed. The
`ALLOWED_HOSTS` setting limits the hosts that Django will serve,[^2] but
we don't know the internal IP in advance, so we can't add it to the
`ALLOWED_HOSTS` setting. Consequently, we set the `Host` header.

Closes #1868

[^1]: https://dokku.com/docs/deployment/zero-downtime-deploys/
[^2]: https://docs.djangoproject.com/en/5.1/ref/settings/#allowed-hosts

Co-authored-by: Steven Maude <git@stevenmaude.co.uk>
Co-authored-by: Katie Bickford <katie.bickford@thedatalab.org>