Use `whitenoise[brotli]` to compress static files.
This is the package that serves static files. We use `whitenoise` for
compression in the `staticfiles` backend in settings.py `STORAGES`. Requiring
brotli will use this more modern and performant compression algorithm rather
than the default gzip. brotli is designed for HTTP and typically has better
compression ratios and decompression performance.