Add custom endpoint for dokku healthchecks
When we deply OpenCodelists, Dokku starts each container and performs a
health check to ensure the container has come up properly before
routing traffic to it. These health checks are customisable (see
[docs](https://dokku.com/docs/deployment/zero-downtime-deploys/)).
Dokku only uses this health check when deploying new containers, it does not use it once the container has been deployed (we do have
Freshping checking the site every few minutes though to make sure it's
still working okay).

This change moves away from using the homepage as the health check
endpoint and to a specific endpoint that will return 200 when the app
starts and nothing else. Using the homepage as a health check is slower
and makes a number of database connections, whereas this in theory
decouples the app health check from the database.

Closes #2154

See https://github.com/opensafely-core/job-server/commit/717946f18