Remove `subprocess` use for management command test.
Using `call_command` is better than `subprocess` because it’s simpler, faster,
more readable, and doesn’t rely on `python` being available. In most cases,
test cases should not use `subprocess` unless testing the integration with
external tools or the shell. This change aligns the test with the rest of the
module and codebase.

This allows us to remove parameters `returncode` (as we now check for an
Exception, not the subprocess return code) and `capfd` (the file descriptor
capture fixture).