Report number of rows imported.
This is useful as a cross-check against the number of mappings in the
spreadsheet.

Updated README to use the command output instead of going looking in the
database manually.

Also added a starting up message as it's not immediately apparent that anything
is happening and it can take a few seconds to run.