Include `format` in `justfile` `check` command.
This ensures that the pre-commit hook that runs `check` will prevent committing
without running `fix` first to get the formatting right. Otherwise you can
commit and push changes then run CI on a PR only to find it has formatting
issues. This happened to me several times and it's annoying! Better to catch
locally before committing. This is in line with what we do in job-server.

Explain difference between `format` and `ruff` in comments.