Take backups via `sqlite ".backup"`.
Taking backups via the sqlite CLI is simple and clean. We can restore from
them.

We haven't been able to locally restore from the `dumpdata` fixtures-based
backups. For now, keep the old-format backups just in case we want to try to
restore again to from them. After 30 days of creating new backups we can update
the script to not consider the old format and location.

Start taking backups in `/storage/backup` rather than directly in `/storage` to
aid navigation and management of backups.

Set -x in `backup.sh` to print the commands used as this aids remote debugging.