Add import_release() for bnfdmd mapping
The "import latest data" management command that
is executed on a scheduled basis expects an
import_release() function. This is implemented
here and unzips the (downloaded, or manually
copied) zip file and calls import_data on the
resultant xlsx file.

Also provided are tests which mock the responses
of the NHSBSA website and ensure that the releases
are correctly parsed and the latest selected.

These mirror the "import latest data" tests for
coding systems.