Pin Django to 4.2 LTS
Until we update the code to work with the latest Django.

The end of life is April 2026, and this is the version we currently use.

This means that update-dependencies-action won't try and update beyond
this minor version. Django is currently on 5.1.x.

From my checking, it's the coding system tests that need updating
so it shouldn't take long to update, but…

…whichever version of Django we choose, it's probably sensible to pin it
regardless, and update when we've tested more thoroughly.