Fix location of DB to take backup from.
Within the container, /app/db.sqlite3 is an apparently empty DB, /storage/db.sqlite3 is the actual app DB of several GB.
I'm not sure why the former exists. But this script was backing that up, not the intended DB.

For example, when this ran in production as scheduled the first time the
produced backup file is only a few tens of bytes in size.

```
mikerkelly@dokku3:~$ ls /var/lib/dokku/data/storage/opencodelists/backup/db/ -lrth
total 4.0K
-rw-r--r-- 1 opencodelists opencodelists 114 Dec 12 01:00 2024-12-12-db.sqlite3.gz
```