Use DATABASE_DIR env var
The path for the core and coding system databases
is set by the DATABASE_DIR environment variable,
which is managed via the dokku config described
in DEPLOY.md.

By using that env var in this backup script,
we ensure that if the configuration of this
variable changes then this script will continue
to work.