Improve testing of import_latest_data for bnfdmd
Clarify name.

Use autospec to ensure mocked function matches
signature of real function. Although this is
increasing coupling between this test and the
import_release function, this test is testing that
this function is called correctly with the right
args and part of that correctness is that the
correct number of arguments are called.

This autospec addition would have caught the bug
introduced in 78b2ada by the omission of the
release_name, valid_from args from
bnfdmd.import_data.import_release and guards their
future removal.