Unit test and fix bnfdmd downloader
download_latest_release() is the entry point for
the Downloader class that is called by the
import_latest_data management command.

This adds unit tests of that entry point given
that the latest mapping file has and has not
already been downloaded.

In case of the latter, the aforementioned
management command expects a ValueError to be
raised, this adds a test for that condition and a
fix such that the test passes.