Remove likely spam users
Management command to apply a series of inclusion
and exclusion heuristics to remove like
non-genuine users.

Users that have contributed to a codelist in some
way, are admins, are members of an organisation,
or have an email address at a trusted domain/TLD
are excluded from deletion.

Users with an email address at a suspicious domain
or TLD, an excessive number of `.` in the local
part of their email address (often used to include
spam phrases), or usernames that don't fit an
expected pattern are included for deletion.

These heuristics were developed in collaboration
with @tomodwyer