Load dotenv in dockerfile
As we no longer rely on environs to do this for us
in settings.py, use the built-in docker env_file
config to load the env vars.

All the docker-based just recipes now depend on
the .env file being present so this also makes
them reliant on the _env recipe which copies from
the sample if not present.

Since the _env just recipe is not called during
production deploy, there will not be a .env file
present in the app root of the container after
deploy and so this change will not change prod's
running configuration.