Consolidate cronfiles and deploy
The three separate cronfiles only differ in the
arg with which the import_latest_release script is
called and their schedules. These can all be
consolidated into a single cronfile with a line
for each invocation.

This simplifies the deployment to /etc/cron.d/ and
allows for this to be scripted and the variables
to be templated, ensuring consistency of
deployment.