Add test helpers for dynamic database-based tests
See #2115 for full details.

The problem this refactoring aims to solve is, in short, as quoted from
@mikerkelly in #2115:

> The [DatabaseOperationForbidden] errors occur because there's a mismatch
> between what [these test cases](https://github.com/opensafely-core/opencodelists/blob/main/coding_systems/bnf/tests/test_import_data.py)
> that import databases dynamically are
> trying to do and how both the Django and `pytest-django` test case
> frameworks work. In a multi-database context, the test frameworks want
> to declare which databases are available to a test case at
> test-case-definition-time, which then triggers those databases being
> created, the test case wrapped in a transaction referencing those
> databases, and the databases flushed in clean up. Since 5.1 it's
> enforced that test cases only open connections to such databases, to
> avoid accidental access. But the [coding system release import tests](https://github.com/opensafely-core/opencodelists/blob/main/coding_systems/bnf/tests/test_import_data.py)
> do want to access a database that they do not want to exist at the start
> of the test. That's in conflict. So maybe we shouldn't be using test cases
> this way at all?

Mike proposed several ideas to fix this that we looked at, and we found that:
amending the test case attribute that defines the allowed databases
seems reasonable, and, further, this was the only solution that we had working.

Altering this attribute is mentioned in Django's documentation:

https://docs.djangoproject.com/en/5.1/topics/testing/tools/#multi-database-support

It was decided that refactoring the existing tests was the solution for
now, rather than rethinking them or the code under test entirely. This
facilitates the upgrade to Django 5.1 without blocking us further. We
may have to revisit this code and these tests in future, with a view to
a more comprehensive overhaul.

The `DynamicDatabaseTestCase class facilitates getting most of the tests
working by just moving them into a test class that uses this
DynamicDatabaseTestCase class. Those tests have a generalisable pattern
for accessing the databases.

The tests that need refactoring fall into one of three general patterns:

* most tests that do not need a temporary coding system database
  directory, handled by the `DynamicDatabaseTestCase` class
* some tests that need a temporary coding system database directory,
  handled by the `DynamicDatabaseTestCaseWithTmpPath` class, which
  extends `DynamicDatabaseTestCase`
* a few edge cases that are discussed in subsequent commits; these will
  be worked around, rather than refactored in this series of commits
  (though they could be reworked in future):
  * tests in `coding_systems/base/tests/test_import_data_utils.py` that
    access the database in fixtures, before the test class' `setUp()`
    runs; these will be handled in that module specifically
  * tests in `coding_systems/bnf/tests/test_import_data.py` that uses a
    pytest fixture to produce import data, instead of fixtures stored in
    the repository, unlike the other tests

Co-authored-by: Mike Kelly <mike.kelly@thedatalab.org>