Reintroduce sentry monitoring of backups
Brings back functionality reverted in 555bc71 but
with three crucial bugfixes:
* correct derivation of script name for sentry
  monitor name
* missing dependency for `curl`
* Fix trailing spaces in cron parsing

Whilst allowing backup script to continue if
sentry logging fails for some reason

The previous code considered a trailing space an
integral part of the cron schedule.

This is erroneous - in the crontab format the
following space is a delimiter for the next field,
and in the app.json it is omitted.

This change fixes the regex such that it correctly
identifies the cron schedule from app.json,
removes the trailing space found in the cronfile,
and updates the tests to (correctly) omit a
trailing space from the test cron schedule.

Refactor backup script with clearer logging,
compared to 555bc72, based on approach taken with
the import script sentry monitoring.

This new approach works within container and is
proven by its use in the import script.