Move shared test case code out from BNF tests
We can reuse most of this code. It's a little messy to do so because of
the use of pytest fixtures; ideally, we sometimes need to reference a
pytest fixture defined in a test module from within our test case class.
Note that not all the existing tests use this feature.