Implement backup sanitisation.
Removes PII (username, name, email) from User table by replacement with
fake data from the Faker library.

Username is used as the PK of User and thus appears in FKs across many
tables, which means these must also be sanitised.

These FKs are configured as "deferrable" which means you are allowed
to break referential integrity in the course of a transaction.

Should any new tables be added that refer to the User table, these will
also be updated by this script.

It is important that the tables referring to the user table are
updated before the user table, otherwise the configured cascade
behaviour of `DELETE` will be invoked.