Allow BNF tests to use dynamic database
These tests mostly fit the general `DynamicDatabaseTestCase` class.

One quirk for these tests is that they use an import data *fixture*
instead of a path to files, unlike the other tests that use import data.

Because of this, there's now a class to handle this case. All the
existing tests use the same fixture.