Add create_user management command
This adds a new `create_user` management command, to aid in easily
creating users/roles in local development and testing.

This is particularly needed following the implementation of backup
sanitisation since production usernames and passwords are sanitised
which prevents developers logging in as their own account.

This is heavily inspired by the management command of the same name
in job server.