Create convenience developer admin user
Since all existing accounts have been sanitised
with fake data, developers own accounts will not
be discoverable in this data.

This management command ensures that there is a
admin developer account with a known password
which has membership of the OpenSAFELY org.

As this creates an admin user with a hard-coded
password, check the settings.DEBUG flag to make
sure this cannot be run in production.

Add to dev-setup recipe after migrations have run
to make sure database is in correct state.