Make `OPTIONS` apply to `DATABASES['default']`.
The `OPTIONS` parameter was a key of `DATABASES`, not `DATABASES["default"]`.
Therefore it didn't apply to any database or take effect.
This can be verified by checking `PRAGMA busy_timeout`, for example in a
management command, as at the bottom of this commit.

This commit fixes that by moving `OPTIONS` into the `default` dict, and using
** keyword unpacking on the `database_url.parse result` to get the other
settings. Now the changed setting is respected, defaulting to 5 seconds as set
by Django (I can't find in the source or docs that SQLite itself applies any
timeout by default, but Django sets a timeout with the PRAGMA, as the
test/command show).

This means that unfortunately #1143, #1811, #2276 had no effect on the timeout.
This might explain why the "database is locked" errors that they sought to
remedy never improved.

This commit does not change the setting from the Django default (for now), so
as to avoid changing the behaviour from what it was before. But this sets up
future commits to change the `OPTIONS` correctly.

Added a test of the setting. I wondered if this is just testing `django`, but
we've seen a minor error in configuring `DATABASES` can lead to unintended
results, so let's test. Also to demonstrate the default does indeed take
effect as claimed here.

---

Management command:
```python
from django.core.management.base import BaseCommand
from django.db import connections

class Command(BaseCommand):
    def handle(self, *args, **options):
        conn = connections["default"]
        raw_conn = conn.connection
        cursor = raw_conn.cursor()
        cursor.execute("PRAGMA busy_timeout")
        timeout_ms = cursor.fetchone()[0]
        print(f"Timeout: {timeout_ms / 1000} seconds")
```