Improve SQLite core DB configuration options.
Options used here are largely inspired by the article
https://kerkour.com/sqlite-for-servers

We think immediate mode in particular might help resolve 'database is locked' errors.

I've largely followed the suggestions from the article, some of which I found
Django applies by default.

I wasn't sure that `PRAGMA temp_store = memory;` was better. This determines
whether files other than the super and rollback journals are written to disk
(eg journal/WAL file of writes to apply; temp index, sort, and join files).
Presumably the main trade off is a marginal speed improvement against possible
additional data loss in some circumstances. E.g., if the process is
interrupted while there are a valid set of pending writes, any incomplete DB
changes can be rolled back and the pending writes may be recoverable and
applied if using file temp_store and consistent with the resulting DB state,
but will be lost if using a memory temp_store. Conversely, maybe the default is
file due to the constrained amount of memory available in the default SQLite
configuration. The differences seem marginal and unrelated to consistency in
any case and I haven't taken time to understand this properly and be
confident in recommending a change.

I did wonder if the very slow search on the OpenCodelists main index page could
be related to using the file system for constructing a large set of joins to
perform the search but that's very speculative. I was also reminded by the
article that "SQLite doesn't keep statistics about its indexes... so COUNT
queries are slow, even when using a WHERE clause on an indexed field: SQLite
has to scan for all the matching records".

I think we should not be using an excessively large timeout as this can
sometimes cause users to wait a long time before getting an error.  Let's try 5
seconds and see if the other settings changes resolve the issue, then gradually
increase the value if there are still many errors.

Previous PRs intended to change the timeout #1143, #1811, #2276 had no effect
as the `DATABASES` property was not correctly structured.