Use VTM as the top of the hierarchy instead of Ing
Having all combination products under a single
hierarchy of an ingredient that matched a search
means that should a user only want a subset of
those combinations they'll have to do a lot of
manual clicking.

VTM is a good proxy for a particular combination
of ingredients, and also combines functionally-
similar ingredients e.g. different salts under
the same VTM.

For these reasons it's more useful than Ing
in the hierarchy.