Run the Python tests with migrations in CI
Fixes #2392.

In #2392, it was specified only to running migrations in CI.

We're not doing things exactly as in job-server here, because job-server
does this outside of Docker, running against a local install (see #2261),
and job-server runs migrations everywhere (opensafely-core/job-server#4573).

It may be that we want to consider running migrations everywhere in future
for OpenCodelists too.

None of that probably matters too much for getting the benefit of adding
extra protection against issues with migrations right now.