Add rudimentary commented-out test
None of this works properly, and is a bit of a throwaway state
ultimately.

But this at least illustrates what Playwright can do with the builder, based
on the codegen tool Playwright provides, and gets you to interacting
with the builder if you run:

```
just test-py --headed
```

On my install at least, with this state, I can get to the builder and
work on a test for the coding system builder (even if the test suite
execution as a whole is all broken, because subsequent tests fail
following the use of Playwright).