Populate development dbs with test fixture data
Closes #2313.

It was suggested in #2313 that using existing test fixture data as dummy data for local development purposes may provide a good alternative to using a copy of prod databases.

The approach taken here shows one way to populate local dbs with test fixture data, based on a comment on the original issue, and a spike (#2407).

Co-authored-by: Jon Massey <jon.massey@thedatalab.org>