Update clean kwargs to suit nh3
`nh3.clean()` does not support the `css_sanitizer` kwarg that `bleach`'s did.

The `MARKDOWN_FILTER_ALLOWLIST_STYLES` setting which fed this was empty so safe to remove.

The `tags` kwarg is required to be a set rather than a list, so corresponding setting updated accordingly.

The `attributes` kwarg has a slightly different API in that nh3 expects a dict of tag:allowed attributes rather than just a list of allowed attributes.
Its possible to replicate the behaviour of bleach (apply the filter to all tags) using "*" as a key.